+++ fuzz/fuzz_targets/shared_library_cranelift.rs
@@ -4,12 +4,12 @@ extern crate libfuzzer_sys;
 
 use wasmer::{imports, Instance, Module, Store};
 use wasmer_compiler_cranelift::Cranelift;
-use wasmer_engine_native::Native;
+use wasmer_engine_shared_library::SharedLibrary;
 
 fuzz_target!(|wasm_bytes: &[u8]| {
     let serialized = {
         let mut compiler = Cranelift::default();
-        let store = Store::new(&Native::new(compiler).engine());
+        let store = Store::new(&SharedLibrary::new(compiler).engine());
         match Module::validate(&store, wasm_bytes) {
             Err(_) => return,
             Ok(_) => {}
@@ -18,7 +18,7 @@ fuzz_target!(|wasm_bytes: &[u8]| {
         module.serialize().unwrap()
     };
 
-    let engine = Native::headless().engine();
+    let engine = SharedLibrary::headless().engine();
     let store = Store::new(&engine);
     let module = unsafe { Module::deserialize(&store, serialized.as_slice()) }.unwrap();
     let _instance = Instance::new(&module, &imports! {});