+++ lib/engine-shared-library/src/lib.rs
@@ -1,9 +1,10 @@
-//! Native backend for Wasmer compilers.
+//! Shared Library Engine for Wasmer compilers.
 //!
 //! Given a compiler (such as `CraneliftCompiler` or `LLVMCompiler`)
-//! it generates a shared object file (`.so` or `.dylib` depending on
-//! the target), saves it temporarily to disk and uses it natively
-//! via `dlopen` and `dlsym` (using the `libloading` library).
+//! this engine compiles a WebAssembly module into a shared object
+//! library (`.so`, `.dylib` or `.dll` depending on the target), saves
+//! it temporarily to disk and uses it natively via `dlopen` and
+//! `dlsym` (using the `libloading` library).
 
 #![deny(missing_docs, trivial_numeric_casts, unused_extern_crates)]
 #![warn(unused_import_braces)]
@@ -27,9 +28,9 @@ mod builder;
 mod engine;
 mod serialize;
 
-pub use crate::artifact::NativeArtifact;
-pub use crate::builder::Native;
-pub use crate::engine::NativeEngine;
+pub use crate::artifact::SharedLibraryArtifact;
+pub use crate::builder::SharedLibrary;
+pub use crate::engine::SharedLibraryEngine;
 
 /// Version number of this crate.
 pub const VERSION: &str = env!("CARGO_PKG_VERSION");